Safari-611.1.10/LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/resource-timing/workerStart-tao-protected.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: undefined is not an object (evaluating 'navigator.serviceWorker.getRegistration')

Harness Error (TIMEOUT), message = null

TIMEOUT Install a SW Test timed out
NOTRUN Add TAO-less iframe to remote origin. Make sure workerStart is zero
NOTRUN Add TAO iframe to remote origin. Make sure workerStart is not zero
